Move MIME check to the only place that handles it
